pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "google-auth-oauthlib>=1.2.2",
    "pillow>=11.2.1",
    "psycopg2-binary>=2.9.10",
    "pypdfium2>=4.30.1",
    "python-dotenv>=1.1.0",
    "requests-oauthlib>=2.0.0",
    "opencv-python>=4.11.0.86",
    "mistralai>=1.7.1",
    "datauri>=1.0.0",
    "azure-ai-documentintelligence>=1.0.0",
]
